ccxt/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
archspec==0.2.3
argcomplete==3.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arviz==0.19.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
backtrader==1.9.78.123
beautifulsoup4==4.12.3
binance-connector==3.8.0
binance-futures-connector==4.0.0
binance.py==1.9.0
bleach==6.1.0
blivet==3.10.0
blivet-gui==2.5.0
boltons==23.1.1
boto3==1.34.162
botocore==1.34.162
branca==0.7.1
Brlapi==0.8.5
build==1.2.1
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chart-studio==1.1.0
cleo==2.1.0
click==8.1.7
cloudpickle==3.0.0
colorlover==0.3.0
comm==0.2.1
conda==24.4.0
conda-libmamba-solver==23.11.1
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
cons==0.4.6
contourpy==1.2.0
crashtest==0.4.1
cryptography==41.0.7
cufflinks==0.17.3
cupshelpers==1.0
cycler==0.12.1
cytoolz==0.12.2
dasbus==1.7
dateparser==1.2.0
dbus-python==1.3.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
distro==1.9.0
dm-tree==0.1.8
dnf==4.21.1
dnspython==2.6.1
dulwich==0.21.7
empiricaldist==0.7.2
etuples==0.3.9
executing==2.0.1
fastjsonschema==2.19.1
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.15.4
folium==0.16.0
fonttools==4.47.0
fqdn==1.5.1
fros==1.1
frozendict==2.3.10
frozenlist==1.4.1
gpg==1.23.2
h11==0.14.0
h5netcdf==1.3.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
humanize==3.13.1
idna==3.7
imageio==2.34.2
install==1.3.5
installer==0.7.0
ipykernel==6.29.0
ipython==8.21.0
ipywidgets==8.1.3
iso639==0.1.4
isoduration==20.11.0
jaraco.classes==3.4.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keyring==24.3.1
kiwisolver==1.4.5
langtable==0.0.68
libcomps==0.1.20
libdnf==0.73.3
libmambapy==1.5.8
llvmlite==0.43.0
logging==0.4.9.6
logical-unification==0.4.6
louis==3.28.0
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
menuinst==2.0.0
miniKanren==1.0.3
mistune==3.0.2
more-itertools==10.2.0
mplfinance==0.12.10b0
msgpack==1.0.8
multidict==6.0.5
multipledispatch==1.0.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nftables==0.1
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.2
olefile==0.47
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pandas_ta==0.3.14b0
pandocfilters==1.5.1
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
pid==2.2.3
pillow==10.2.0
pkginfo==1.10.0
platformdirs==3.11.0
plotly==5.22.0
pluggy==1.3.0
ply==3.11
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
polars==0.20.31
productmd==1.38
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pwquality==1.4.5
py-cpuinfo==9.0.0
pyAgrum==1.15.0
pycairo==1.25.1
pycosat==0.6.6
pycparser==2.20
pycryptodome==3.20.0
pycups==2.0.4
pycurl==7.45.2
pydot==3.0.1
pyenchant==3.2.2
pyfolio==0.9.2
pygame==2.5.2
pygdbmi==0.11.0.0
Pygments==2.17.2
PyGObject==3.48.2
pykickstart==3.52
pymc==5.16.2
Pyomo==6.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.2
pyparted==3.13.0
pyproject_hooks==1.0.0
PySocks==1.7.1
pytensor==2.25.2
python-augeas==1.1.0
python-binance==1.0.19
python-dateutil==2.8.2
python-json-logger==2.0.7
python-meh==0.51
python-pam==2.0.2
pytz==2023.3.post1
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.8.1
ratelimiter==1.2.0.post0
referencing==0.35.1
regex==2024.4.16
requests==2.31.0
requests-file==2.0.0
requests-ftp==0.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rpm==4.19.1.1
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.7
s3transfer==0.10.2
schedule==1.2.2
scikit-learn==1.5.1
scipy==1.11.4
seaborn==0.13.2
SecretStorage==3.3.3
selinux @ file:///builddir/build/BUILD/libselinux-3.6/src
Send2Trash==1.8.3
sentry-sdk==2.11.0
sepolicy @ file:///builddir/build/BUILD/selinux-3.6/python/sepolicy
setools==4.5.1
setproctitle==1.2.3
setuptools==69.0.3
shellingham==1.5.4
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
sos==4.7.2
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.2
systemd-python==235
tenacity==8.4.2
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tomlkit==0.12.4
toolz==0.12.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
trove-classifiers==2024.3.25
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2023.4
tzlocal==5.2
ujson==5.10.0
uri-template==1.3.0
urllib3==1.26.19
utils==1.0.2
vectorbt==0.26.1
virtualenv==20.25.1
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
widgetsnbextension==4.0.11
xarray==2024.6.0
xarray-einstats==0.7.0
xkbregistry==0.3
xyzservices==2023.10.1
yarl==1.9.4
zstandard==0.22.0
